Implemented add to cart button functionality and cart screen
Created two lists to store items in user's cart and they get rendered in the cart screen with their image, name, qty, and price. One for item objects and the other with item names to prevent multiple copies of item objects being rendered when adding more quantity to an exisitng item, it simply increments it quantity instead.